MYNEWT-347 BLE Host - GAP update; use L2CAP proc.
Prior to this change, the host did not failover to the L2CAP connection
parameter update procedure when the GAP update procedure failed.
Rather, it was up to the application to manually do that.

Now, if the host automatically performs the L2CAP procedure if all of
the following are true:
    * We are the slave.
    * Controller reports one of the following status codes in its
      connection update complete event:
        o Unknown HCI command (0x01).
        o Unsupported remote feature (0x1a).

In addition, there was some missing functionality that this commit fills
in:
    * Timeouts for GAP update procedures (30 seconds).
    * Report and delete outstanding L2CAP sig procedures when the
      connection is broken.